chore: :adhesive_bandage: Fix norminette
